events and localstorage testing
Was thinking that storing the objects in localstorage would allow for html element object linking, but that doesn't seem to be the case. Will have to figure out a way to link the pieces on the board to their respective objects. At least I am able to save the object names within each image tag so that might be helpful.