no pawn immediate checkmate implemented
had a version of this with too many function calls to checkAllPiecesForKingCheckMate(). Made a new function to first check if either king is in the pawn drop's possible moves. If either are then CheckAllPiecesForKingCheckMate() is called.

To increase efficiency might make checkIfKingInCheck() used first instead of checkAllPiecesForKingCheck() and use checkAllPiecesForKingCheck() only when very nessisary, unsure if possible though since checkAllPiecesForKingCheck() was made to solve previous bugs such as pieces being able to move out of protecting the king.

A pawn may not be dropped to give an immediate checkmate

Finally all that's left are some edits to when click events are put and disabled (before the other player joins and when asking to promote), a bunch of socketio, and styling, then finally release!!!!!!